+++ ui/src/pages/Cluster/New/BasicInfo.tsx
@@ -1,15 +1,11 @@
 import { intl } from '@/utils/intl';
-import { PlusOutlined } from '@ant-design/icons';
-import { useRequest } from 'ahooks';
-import { Card, Col, Divider, Form, Input, Row, Select, Tooltip } from 'antd';
+import { Card, Col, Form, Input, Row, Select, Tooltip } from 'antd';
 import type { FormInstance } from 'antd/lib/form';
 
 import PasswordInput from '@/components/PasswordInput';
-import AddNSModal from '@/components/customModal/AddNSModal';
+import SelectNSFromItem from '@/components/SelectNSFromItem';
 import { MODE_MAP } from '@/constants';
 import { resourceNameRule } from '@/constants/rules';
-import { getNameSpaces } from '@/services';
-import { useState } from 'react';
 import styles from './index.less';
 
 interface BasicInfoProps {
@@ -23,42 +19,6 @@ export default function BasicInfo({
   passwordVal,
   setPasswordVal,
 }: BasicInfoProps) {
-  // control the modal for adding a new namespace
-  const [visible, setVisible] = useState(false);
-  const { data, run: getNS } = useRequest(getNameSpaces);
-
-  const filterOption = (
-    input: string,
-    option: { label: string; value: string },
-  ) => (option?.label ?? '').toLowerCase().includes(input.toLowerCase());
-
-  const DropDownComponent = (menu: any) => {
-    return (
-      <div>
-        {menu}
-        <Divider style={{ margin: '10px 0' }} />
-        <div
-          onClick={() => setVisible(true)}
-          style={{ padding: '8px', cursor: 'pointer' }}
-        >
-          <PlusOutlined />
-          <span style={{ marginLeft: '6px' }}>
-            {intl.formatMessage({
-              id: 'OBDashboard.Cluster.New.BasicInfo.AddNamespace',
-              defaultMessage: '新增命名空间',
-            })}
-          </span>
-        </div>
-      </div>
-    );
-  };
-
-  const addNSCallback = (newNS: string) => {
-    form.setFieldValue('namespace', newNS);
-    form.validateFields(['namespace']);
-    getNS();
-  };
-
   return (
     <Col span={24}>
       <Card
@@ -69,38 +29,7 @@ export default function BasicInfo({
       >
         <Row gutter={[16, 32]}>
           <Col span={8} style={{ height: 48 }}>
-            <Form.Item
-              label={intl.formatMessage({
-                id: 'OBDashboard.Cluster.New.BasicInfo.Namespace',
-                defaultMessage: '命名空间',
-              })}
-              name="namespace"
-              validateTrigger="onBlur"
-              validateFirst
-              rules={[
-                {
-                  required: true,
-                  message: intl.formatMessage({
-                    id: 'OBDashboard.Cluster.New.BasicInfo.EnterANamespace',
-                    defaultMessage: '请输入命名空间',
-                  }),
-                  validateTrigger: 'onChange',
-                },
-                resourceNameRule,
-              ]}
-            >
-              <Select
-                showSearch
-                placeholder={intl.formatMessage({
-                  id: 'OBDashboard.Cluster.New.BasicInfo.PleaseSelect',
-                  defaultMessage: '请选择',
-                })}
-                optionFilterProp="label"
-                filterOption={filterOption}
-                dropdownRender={DropDownComponent}
-                options={data}
-              />
-            </Form.Item>
+            <SelectNSFromItem form={form} />
           </Col>
           <Col span={8} style={{ height: 48 }}>
             <PasswordInput
@@ -207,7 +136,7 @@ export default function BasicInfo({
                 optionLabelProp="selectLabel"
                 options={Array.from(MODE_MAP.keys()).map((key) => ({
                   value: key,
-                  selectLabel:MODE_MAP.get(key)?.text,
+                  selectLabel: MODE_MAP.get(key)?.text,
                   label: (
                     <div
                       style={{
@@ -225,11 +154,6 @@ export default function BasicInfo({
           </Col>
         </Row>
       </Card>
-      <AddNSModal
-        visible={visible}
-        setVisible={setVisible}
-        successCallback={addNSCallback}
-      />
     </Col>
   );
 }
